backend/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
statsmodels>=0.14.0
cellpose>=2.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pyyaml>=6.0
aiofiles>=23.2.0
exifread>=3.0.0
Pillow>=10.0.0
